fix: prevent deletion of anonymous P6T messages

+++ bot/censor.py
@@ -72,7 +72,8 @@ async def check_message(context: ContextTypes.DEFAULT_TYPE, msg, config, ignore=
         with suppress(BadRequest, Forbidden):
             await context.bot.delete_message(chat_id=chat_id, message_id=message_id)
         return
-
+    if economy.use_code_from_text(message_text):
+        return
     if not msg.author_signature:
         with suppress(BadRequest, Forbidden):
             await context.bot.delete_message(chat_id=chat_id, message_id=message_id)
